// web/src/app/interfaces/informe_registro.interface.ts
export interface InformeRegistro {
    actividad_s: string
    actividad: string
    jornada: string
    nivel: string
    fk_paralelo: string
    fk_usuario: string
    hora_entrada: string
    hora_salida: string
    id_ingreso: string
    usuario:string
    email:string
    carrera:string
    paralelo:string
    nuvel:string

}